Complete AD19 read mutation output consistency repair
