+++ docs/requirements.md
@@ -1230,6 +1230,13 @@ Required behavior:
 - persist read-triggered state changes back to the physical inbox file that
   owns the selected displayed message when origin inbox files are present in
   the merged surface
+- when a read-side mutation is applied, the returned `message` payload and
+  `selected_message_id` must still refer to that same mutated durable message;
+  `atm read` must not mark one message read and then silently swap the output
+  payload to a different unread message
+- `bucket_counts` in the read outcome must describe the post-mutation mailbox
+  state produced by that command execution rather than stale pre-mutation
+  counts
 
 ### 7.5 Shared Message Classification And Deduplication
 
@@ -1387,6 +1394,15 @@ Every list row must include:
 - `match_count`
 - `additional_match_count`
 - `bucket_counts`
+
+When `mutation_applied = true` and `message` is present:
+- `message.message_id` and `selected_message_id` must identify the same
+  durable message
+- `bucket_counts` must reflect the mailbox state after the read-side mutation
+  completes
+- the read-side mutation contract is distinct from `atm ack`; read may mark a
+  message `read = true`, but only ack clears `pending_ack_at` and sets
+  `acknowledged_at`
 
 Human-readable `atm read` output must render one message body only. When
 additional matches exist, it must state that more matches were found and direct
